services/vision/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
pydantic==2.12.3
python-multipart==0.0.20
torch==2.9.0
torchvision==0.24.0
Pillow==12.0.0
numpy==2.3.4
